extract TreeNodeContainer to its own compilation unit
