+++ MSGViewer/src/main/java/net/sourceforge/MSGViewer/MSGNavigator/EditNode.java
@@ -3,7 +3,6 @@ package net.sourceforge.MSGViewer.MSGNavigator;
 import at.redeye.FrameWork.base.AutoMBox;
 import at.redeye.FrameWork.base.BaseDialog;
 import at.redeye.FrameWork.base.Root;
-import net.sourceforge.MSGViewer.MSGNavigator.MSGNavigator.TreeNodeContainer;
 import org.apache.poi.poifs.filesystem.DirectoryEntry;
 import org.apache.poi.poifs.filesystem.DocumentEntry;
 import org.apache.poi.poifs.filesystem.Entry;
